refactor: Introducing the ConfigurationHelper idea
