+++ alert-common/src/main/java/com/synopsys/integration/alert/common/rest/api/ConfigurationHelper.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.alert.common;
+package com.synopsys.integration.alert.common.rest.api;
 
 import java.util.Optional;
 import java.util.function.Supplier;
